PUBDEV-5378: Moved introduced logic from Frame.java to TargetEncodingFrameHelper.java
