+++ h2o-automl/src/test/java/ai/h2o/automl/targetencoding/AllTETestsSuite.java
@@ -1,9 +1,11 @@
 package ai.h2o.automl.targetencoding;
 
 
+import org.junit.Ignore;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
+@Ignore // just for the case if Jenkins is picking up this class causing tests to run two times
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
         TargetEncodingTest.class,
@@ -13,6 +15,7 @@ import org.junit.runners.Suite;
         TargetEncodingMultiClassTargetTest.class,
         TargetEncodingTargetColumnTest.class,
         TargetEncodingExceptionsHandlingTest.class,
+        TargetEncodingFrameHelperTest.class,
         TargetEncodingImmutabilityTest.class
 })
