+++ h2o-core/src/main/java/water/rapids/ast/prims/advmath/AstUnique.java
@@ -1,9 +1,15 @@
 package water.rapids.ast.prims.advmath;
 
+import water.DKV;
+import water.MRTask;
+import water.fvec.Chunk;
 import water.fvec.Frame;
+import water.fvec.Vec;
+import water.fvec.task.UniqTask;
 import water.rapids.Env;
 import water.rapids.ast.AstPrimitive;
 import water.rapids.ast.AstRoot;
+import water.rapids.ast.prims.mungers.AstGroup;
 import water.rapids.vals.ValFrame;
 
 public class AstUnique extends AstPrimitive {
@@ -25,7 +31,30 @@ public class AstUnique extends AstPrimitive {
   @Override
   public ValFrame apply(Env env, Env.StackHelp stk, AstRoot asts[]) {
     Frame fr = stk.track(asts[1].exec(env)).getFrame();
-    return new ValFrame(fr.uniqueValuesBy(0));
+    return new ValFrame(uniqueValuesBy(fr,0));
   }
 
+  /** return a frame with unique values from the specified column */
+  public static Frame uniqueValuesBy(Frame fr, int columnIndex) {
+    Vec vec0 = fr.vec(columnIndex);
+    Vec v;
+    if (vec0.isCategorical()) {
+      v = Vec.makeSeq(0, (long) vec0.domain().length, true);
+      v.setDomain(vec0.domain());
+      DKV.put(v);
+    } else {
+      UniqTask t = new UniqTask().doAll(vec0);
+      int nUniq = t._uniq.size();
+      final AstGroup.G[] uniq = t._uniq.keySet().toArray(new AstGroup.G[nUniq]);
+      v = Vec.makeZero(nUniq, vec0.get_type());
+      new MRTask() {
+        @Override
+        public void map(Chunk c) {
+          int start = (int) c.start();
+          for (int i = 0; i < c._len; ++i) c.set(i, uniq[i + start]._gs[0]);
+        }
+      }.doAll(v);
+    }
+    return new Frame(v);
+  }
 }
